requirements.txt
tensorflow-macos==2.13.0
tensorflow-model-optimization
numpy
matplotlib
